Working on repeatable random number generation.
Replacing all uses of Math.random() with and instance of the Random
class that can be seeded. Adding passing of long seeds or Random
instances among classes, and adding an option to the main function to
specify a seed.

Nothing has been tested yet, and things are likely broken at this point.